Added the LM566 to our MAME library.
This chip has been drawn in the "proper" way, meaning that it looks exactly like
it should in a schematic sheet (not the hard-and-fast DIP package orientation).

I would like to go back and fix rest of the chips to look like this as well, but
we'll have to see how things work out.